refactor: change the directory structure #1
